+++ previous-versions/sync-for-expenses-version-1/pkg/models/operations/deleteconnection.go
@@ -8,7 +8,9 @@ import (
 )
 
 type DeleteConnectionRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 }
 
@@ -27,11 +29,14 @@ func (o *DeleteConnectionRequest) GetConnectionID() string {
 }
 
 type DeleteConnectionResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *DeleteConnectionResponse) GetContentType() string {