+++ previous-versions/sync-for-expenses-version-1/pkg/models/operations/getsyncbyid.go
@@ -8,6 +8,7 @@ import (
 )
 
 type GetSyncByIDRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 	// Unique identifier for a sync.
 	SyncID string `pathParam:"style=simple,explode=false,name=syncId"`
@@ -30,11 +31,14 @@ func (o *GetSyncByIDRequest) GetSyncID() string {
 type GetSyncByIDResponse struct {
 	// Success
 	CompanySyncStatus *shared.CompanySyncStatus
-	ContentType       string
+	// HTTP response content type for this operation
+	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetSyncByIDResponse) GetCompanySyncStatus() *shared.CompanySyncStatus {