+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [go v0.24.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [Go v0.24.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.24.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-09-26 00:04:35
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.91.0 (2.129.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.25.0 - previous-versions/sync-for-expenses-version-1